fix to show image on web app once uploaded
Passing the Blob URL to the Template:

In app.py, the upload_form function now passes the filename (which is actually the blob URL) to the template.
Rendering the Image in the Template:

In upload.html, the image is displayed using the img tag with src set to {{ filename }} if filename is available.